De-duplicate table commands code.
