+++ src/tableproperties/commands/tablealignmentcommand.js
@@ -7,9 +7,7 @@
  * @module table/tableproperties/commands/tablealignmentcommand
  */
 
-import Command from '@ckeditor/ckeditor5-core/src/command';
-
-import { findAncestor } from '../../commands/utils';
+import TablePropertyCommand from './tablepropertycommand';
 
 /**
  * The table alignment command.
@@ -20,62 +18,18 @@ import { findAncestor } from '../../commands/utils';
  * To change alignment of the selected table, execute the command:
  *
  *		editor.execute( 'tableAlignment', {
- *			value: '5px'
+ *			value: 'right'
  *		} );
  *
- * @extends module:core/command~Command
+ * @extends module:table/tableproperties/commands/tablepropertycommand
  */
-export default class TableAlignmentCommand extends Command {
-	constructor( editor ) {
-		super( editor );
-
-		this.attributeName = 'alignment';
-	}
-
+export default class TableAlignmentCommand extends TablePropertyCommand {
 	/**
-	 * @inheritDoc
-	 */
-	refresh() {
-		const editor = this.editor;
-		const selection = editor.model.document.selection;
-
-		const table = findAncestor( 'table', selection.getFirstPosition() );
-
-		this.isEnabled = !!table;
-		this.value = this._getValue( table );
-	}
-
-	_getValue( table ) {
-		if ( !table ) {
-			return;
-		}
-
-		return table.getAttribute( this.attributeName );
-	}
-
-	/**
-	 * Executes the command.
+	 * Creates a new `TableAlignmentCommand` instance.
 	 *
-	 * @fires execute
-	 * @param {Object} [options]
-	 * @param {Boolean} [options.value] If set the command will set alignment.
-	 * If alignment is not set the command will remove the attribute.
+	 * @param {module:core/editor/editor~Editor} editor Editor on which this command will be used.
 	 */
-	execute( options = {} ) {
-		const model = this.editor.model;
-		const selection = model.document.selection;
-
-		const { value, batch } = options;
-
-		const tables = Array.from( selection.getSelectedBlocks() )
-			.map( element => findAncestor( 'table', model.createPositionAt( element, 0 ) ) );
-
-		model.enqueueChange( batch || 'default', writer => {
-			if ( value ) {
-				tables.forEach( table => writer.setAttribute( this.attributeName, value, table ) );
-			} else {
-				tables.forEach( table => writer.removeAttribute( this.attributeName, table ) );
-			}
-		} );
+	constructor( editor ) {
+		super( editor, 'alignment' );
 	}
 }
