+++ src/tableproperties/commands/tableheightcommand.js
@@ -7,9 +7,7 @@
  * @module table/tableproperties/commands/tableheightcommand
  */
 
-import Command from '@ckeditor/ckeditor5-core/src/command';
-
-import { findAncestor } from '../../commands/utils';
+import TablePropertyCommand from './tablepropertycommand';
 
 /**
  * The table height command.
@@ -20,61 +18,18 @@ import { findAncestor } from '../../commands/utils';
  * To change height of the selected table, execute the command:
  *
  *		editor.execute( 'tableHeight', {
- *			value: '5px'
+ *			value: '500px'
  *		} );
  *
  * @extends module:core/command~Command
  */
-export default class TableHeightCommand extends Command {
-	constructor( editor ) {
-		super( editor );
-
-		this.attributeName = 'height';
-	}
-
+export default class TableHeightCommand extends TablePropertyCommand {
 	/**
-	 * @inheritDoc
-	 */
-	refresh() {
-		const editor = this.editor;
-		const selection = editor.model.document.selection;
-
-		const table = findAncestor( 'table', selection.getFirstPosition() );
-
-		this.isEnabled = !!table;
-		this.value = this._getValue( table );
-	}
-
-	_getValue( table ) {
-		if ( !table ) {
-			return;
-		}
-
-		return table.getAttribute( this.attributeName );
-	}
-
-	/**
-	 * Executes the command.
+	 * Creates a new `TableHeightCommand` instance.
 	 *
-	 * @fires execute
-	 * @param {Object} [options]
-	 * @param {Boolean} [options.value] If set the command will set height. If height is not set the command will remove the attribute.
+	 * @param {module:core/editor/editor~Editor} editor Editor on which this command will be used.
 	 */
-	execute( options = {} ) {
-		const model = this.editor.model;
-		const selection = model.document.selection;
-
-		const { value, batch } = options;
-
-		const tables = Array.from( selection.getSelectedBlocks() )
-			.map( element => findAncestor( 'table', model.createPositionAt( element, 0 ) ) );
-
-		model.enqueueChange( batch || 'default', writer => {
-			if ( value ) {
-				tables.forEach( table => writer.setAttribute( this.attributeName, value, table ) );
-			} else {
-				tables.forEach( table => writer.removeAttribute( this.attributeName, table ) );
-			}
-		} );
+	constructor( editor ) {
+		super( editor, 'height' );
 	}
 }
